requirements.txt
Django==3.1
psycopg2-binary==2.8.5
Pillow==8.0
djangorestframework==3.11.0
python-decouple==3.4
django-phonenumber-field==5.0.0
phonenumbers==8.12.19
djangorestframework-simplejwt
drf-yasg
celery
redis
beautifulsoup4==4.9.3
lxml==4.6.2
requests==2.25.1
